Took out some warnings when compiling and added Key puzzle.

+++ void_rush/src/generation/Generation_manager.cpp
@@ -1,7 +1,7 @@
 #include "Generation_manager.hpp"
 
 Generation_manager::Generation_manager(Graphics*& _gfx, ResourceManager*& _rm, CollisionHandler& collisionHandler)
-	: gfx(_gfx), rm(_rm), seed(time(0)), difficulity(Difficulity::easy)
+	: gfx(_gfx), rm(_rm), seed((int)time(0)), difficulity(Difficulity::easy)
 {        
     
     this->position_gen = new Position_generator(this->seed);
@@ -81,11 +81,11 @@ void Generation_manager::generateGraph()
     out << "digraph {\n";
     int scale = 20;
 
-    int abs_X = 0;
-    int abs_Y = 0;
+    float abs_X = 0;
+    float abs_Y = 0;
     Player_jump_checker p;
     std::vector<Platform*>* platforms = this->position_gen->getPlatforms();
-    int platforms_size = platforms->size();
+    int platforms_size = (int)platforms->size();
     for (int i = 0; i < platforms_size; i++)
     {
